Make 3. maongodb node as delayed back-up

+++ infra/mongodb/clean.sh
@@ -1,5 +1,15 @@
 #!/bin/bash
 
+# Force run as root
+if [ $(id -u) != "0" ]
+    then
+        sudo "$0" "$@"  # Modified as suggested below.
+        exit $?
+fi
+
+# Kill all running mongod processes
+pkill -9 mongod
+
 rm -Rf /data-mongodb/xuser-1
 rm -Rf /data-mongodb/xuser-2
 rm -Rf /data-mongodb/backup
